feat(offline flere brukere): legg til støtte for flere brukere og andre
Lagde sessions tabell også for å få legge til logout funksjon senere

Forbedre log filens feil melding for hjemside endepunkt

Skal ikke være online. POST /signup endepunkt gir feil meldingen: "Data handling failed (user signup)! Table 'notes_app.users'
doesn't exist"

Andre
Feil rett for /patch-todo endepunkt (funksjon patch_row) - ved et uhell

+++ modules/delete_row.py
@@ -3,6 +3,6 @@ from typing import Literal
 T_To_Delete = Literal["Users", "Notes", "TODOs"]
 
 
-def delete_row(cursor, t_to_delete: T_To_Delete, row_id: int) -> None:
-    delete_query = f"DELETE FROM {t_to_delete} WHERE id = ?"
-    cursor.execute(delete_query, [row_id])
+def delete_row(cursor, t_to_delete: T_To_Delete, row_id: int, user_id: int) -> None:
+    delete_query = f"DELETE FROM {t_to_delete} WHERE id = ? AND user_id = ?"
+    cursor.execute(delete_query, (row_id, user_id))
